+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## 2.0.0 (unpublished)
 
 KeePass 2.47 or newer is required now (was 2.43).
-
 Micrososft .Net 4.6.1 is required now (was 4.5.2).
+Amazon Drive support has been dropped.
 
 ### Fixed
 
+- \#246 Remove Amazon Drive support
 - \#241 \[Google Drive\] Cannot Authorize (browser not supported) bug 
 - \#227 "Couldn't sign you in" GDrive error bug 
 - \#212 Can't connect to google drive account